refactor and create ChallengeAndAnswer record
