+++ backend/src/main/java/org/josh/backend/dto/ChallengeAndAnswer.java
@@ -1,0 +1,7 @@
+package org.josh.backend.dto;
+
+public record ChallengeAndAnswer(
+    String challenge,
+    String answer
+) {
+}
