+++ backend/src/main/java/org/josh/backend/openai/PromptBuilder.java
@@ -1,6 +1,8 @@
 package org.josh.backend.openai;
 
-import org.josh.backend.workshop.WorkshopFormData;
+import org.josh.backend.dto.Gpt3TurboRequest;
+import org.josh.backend.dto.Gpt3TurboResponse;
+import org.josh.backend.dto.WorkshopFormData;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
