+++ backend/src/main/java/org/josh/backend/workshop/WorkshopController.java
@@ -1,8 +1,11 @@
 package org.josh.backend.workshop;
 
 import lombok.RequiredArgsConstructor;
+import org.josh.backend.dto.WorkshopFormData;
 import org.josh.backend.exception.ErrorMessage;
 import org.josh.backend.exception.NoSuchWorkshopException;
+import org.josh.backend.dto.ChallengeAndAnswer;
+import org.josh.backend.dto.Gpt3TurboResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -38,6 +41,11 @@ public class WorkshopController {
     @DeleteMapping("/{id}")
     public void deleteWorkshop(@PathVariable String id) {
         workshopService.deleteWorkshop(id);
+    }
+
+    @PostMapping("/{id}/validate")
+    public Gpt3TurboResponse validateChallenge(@PathVariable String id, @RequestBody ChallengeAndAnswer challengeAndAnswer) {
+        return workshopService.validateChallenge(id, challengeAndAnswer);
     }
 
     @ExceptionHandler({NoSuchWorkshopException.class})
